refactor: replace ColoredPrinter with single impl
Since Unix works for all OS, there's no point in one additional indirection. I can think of any other implementation (it's not like with Printer, where you can have Terminal and File printers). Colors only show on Terminals.
